requirements.txt
streamlit==1.29.0
pypdf2==3.0.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.35.0
torch==2.1.0
protobuf==3.20.3
